cryptoauthlib: use digi-embedded github mirror
The commit we use to build the cryotpauthlib package is currently not part of
any branches or tags in the original MicrochipTech repo. To make sure the
package can always be built, use our mirror of the repo, which includes an
additional branch for the specific commit.

While at it, move the "nobranch=1" parameter to the GIT_URI variable, since
we're putting all other git parameters there, anyway. Also, remove the "branch"
parameter because it isn't needed in this case.

https://onedigi.atlassian.net/browse/DEL-8015

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>