packagegroup-x-linux-ai: remove c++ examples on the ccmp13
These examples have a gtk+3 dependency that can't be met on the ccmp13 due to
it being a headless device. Remove these examples from the packagegroups so we
can at least build the rest of the ML packages.

Note that all of ST's ML examples are GUI-based and will not be usable on the
ccmp13, but since the remaining examples don't have an explicit gtk+3
dependency, at least they won't trigger build errors when included in the
image.

https://onedigi.atlassian.net/browse/DEL-8308

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>