packagegroup-qt6-dey: remove qtquick3d-dev
This package was inadvertantly pulling in a lot of dependencies into our
images that make use of Qt6. Most of these packages consisted of *-dev
packages, which are only useful for development with the SDK.

Remove this package along with its dependencies to significantly reduce the
dey-image-qt rootfs image size.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>