meta-digi-dey: python3-xbee: update recipe to compile Github code
Instead of depending on library releases in Pypi, update the recipe to compile latest
Github code. This allows to use minor fixes in DEY before a new release of the library
is available. While on it, set the 'SRCREV' to point to the last commit of the repository
for traceability.

Signed-off-by: David Escalona <david.escalona@digi.com>