imx-gpu-viv: avoid pulling in kernel-module-imx-gpu-viv
We used to use BAD_RECOMMENDATIONS to remove this package in ccimx6 builds,
we enable the imx-gpu-viv driver as built-in in our kernel, but this method
isn't working anymore. Instead, undo the specific RRECOMMENDS that pulls the
module in.

Apply the change for the aarch32 version of the package only, since this change
is only needed for the ccimx6 platforms.

https://onedigi.atlassian.net/browse/DEL-8540

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>