Add meta-multimedia layer for ccimx8/ccimx9 platforms
NXP's machine learning packagegroup now has a dependency for gst-shark, a GPU
profiling tool which is provided by this layer. Since it's a part of a repo
that already exists in our manifest, simply add it to the default layers of all
i.MX platforms that support the machine learning feature to be able to include
its packages out of the box.

https://onedigi.atlassian.net/browse/DEL-8551

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>