u-boot-dey: re-order ccimx8x deployment steps to avoid errors
For the ccimx8x, we changed the order of the steps in do_deploy() from:

    Deploy -> Rename files -> Move binaries

To:

    Deploy -> Move binaries -> Rename files

When it's time to rename the files, they won't be in their original place and
the process will fail. Make sure we move the files after they've been renamed
to avoid errors.

Also, one move operation is enough for all artifacts, so remove the second
operation.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>