Build U-Boot v2020.04, qca driver and Linux v5.15 from cc8x-cc6-beta branches
These branches contain the necessary changes to support the ccimx8x and ccimx6
platforms in the DEY 4.0 beta release. Note that we don't need a new branch for
U-Boot v2017.03 (the version used for the ccimx6) becase there hasn't been any
new development in that branch since its last release.

https://onedigi.atlassian.net/browse/DEL-8653

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>