packagegroup-qt6-dey: remove qtquick3d-dev
This package was inadvertantly pulling in a lot of dependencies into our
images that make use of Qt6. Most of these packages consisted of *-dev
packages, which are only useful for development with the SDK.

Remove this package along with its dependencies to significantly reduce the
dey-image-qt rootfs image size.

Note that this change was already done in commit
8979331ef8a2a47e2368d33ae58c36f73f937d33, but it was accidentally reverted when
upgrading to Qt 6.5 in commit a5c6f2b795467b2cb3d21e73c911638d227e3a3e.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>