tf-a-stm32mp: create symlinks on deploydir to TF-A and FIP binaries
These binaries are installed in subdirectories by default. The uuu
installer expects to find all binaries on the same folder where the script
is. By creating symlinks, the uuu installer can find all the binaries it
needs directly on the deploy folder.

NOTE: variables in 'for' clauses are intentionally without quotes to skip
whitespaces in them.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>