udev: automount: add exception to mount loop and ram block devices
The nodes "/dev/ramX" and "/dev/loopX" are mounted on boot.
Each node calls the mount.sh script, but they are not mounted
because these nodes are blacklisted in the "blacklist.conf" file.

In the ccmp13 adding this modification in the rule saves
around 4 seconds per boot.
In the ccmp15 and ccimx6ul around 2 seconds are saved.

https://onedigi.atlassian.net/browse/DEL-8725

Signed-off-by: Francisco Gil <francisco.gilmartinez@digi.com>