uboot: ccmp1: disable LEGACY_IMAGE_FORMAT when FIT enabled
The LEGACY_IMAGE_FORMAT defaults to 'y' if there is no FIT
support, which happens after applying the default configuration.
Then, the FIT support is added in a config fragment, but the
LEGACY_IMAGE_FORMAT is not disabled.
Disabling this is recommended to avoid the possibility to boot
unsigned legacy images.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>