trustfence-sign-tools: new recipe based on u-boot v2023.04
The ccimx93 bootloader signing requires this version of the sign script,
but at the moment, the updated script is not compatible with the other
AHAB-based module (ccimx8x), so we cannot just update the recipe for
all modules.

Thus, create a new version of the recipe and use that for the ccimx93
while we still use the old version for the rest of the platforms.

https://onedigi.atlassian.net/browse/DUB-1068

Signed-off-by: Javier Viguera <javier.viguera@digi.com>