linux-virtualization: expand docker config extract
expand the docker defconfig excerpt to add more default options, as some
of them might be enable in some platform defconfigs but not in other ones,
so just set all of them, as it is safe, and nothing happens if they are
already set in the original default defconfig.

To check if all LXC/docker options are enabled for a kernel,
run lxc-checkconfig in the system.

https://onedigi.atlassian.net/browse/DEL-8924

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>