build.sh: fix error when applying CVE layer
In our previous commit, we changed the CVE scan class from "vigiles" to
"digi_ccss" if we plan on building images with the CVE layer. However, we
make this change in conf/local.conf and then run "bitbake-layers add-layer" to
add said layer. Since the bbclass is exclusive to the CVE layer, bitbake isn't
able to recognize it and fails.

Add the CVE layer to the project before adding the Vigiles configuration
template to conf/local.conf.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>